Rename "registry" file to "registries"
